+++ src/SMAPI.Web/Framework/LogParsing/Models/ParsedLog.cs
@@ -25,6 +25,9 @@ namespace StardewModdingAPI.Web.Framework.LogParsing.Models
         /****
         ** Log data
         ****/
+        /// <summary>Whether SMAPI is running in strict mode, which disables all deprecated APIs.</summary>
+        public bool IsStrictMode { get; set; }
+
         /// <summary>The SMAPI version.</summary>
         public string? ApiVersion { get; set; }